cpython: Un-remove some test data
These certificates are used in asyncio's tests in a way that doesn't
affect CPython directly, but does affect some dependencies, in
particular txaio.

Fixes NixOS/nixpkgs#53474.

+++ pkgs/development/interpreters/python/cpython/3.5/default.nix
@@ -133,6 +133,7 @@ in stdenv.mkDerivation {
     for item in $out/lib/python${majorVersion}/test/*; do
       if [[ "$item" != */test_support.py*
          && "$item" != */test/support
+         && "$item" != */test/*.pem
          && "$item" != */test/libregrtest
          && "$item" != */test/regrtest.py* ]]; then
         rm -rf "$item"
